% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/testUtils.R
\name{lclapply}
\alias{lclapply}
\title{Calls mclapply and combines the end result using listcomb}
\usage{
lclapply(x, f, mc.cores = 0)
}
\description{
Calls mclapply and combines the end result using listcomb
}
\examples{
lclapply(1:10, function(i) list(a = i, b = i^2), mc.cores = 1)
}

