% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/testUtils.R
\name{listcomb}
\alias{listcomb}
\title{Converts list(list(fields)) into list(fields(list))}
\usage{
listcomb(dots)
}
\arguments{
\item{dots}{list of lists}
}
\description{
Converts list(list(fields)) into list(fields(list))
}

