% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/testUtils.R
\name{mclapply0}
\alias{mclapply0}
\title{Either uses lapply or mclapply}
\usage{
mclapply0(x, f, mc.cores = 0)
}
\description{
Either uses lapply or mclapply
}
\examples{
lclapply(1:10, function(i) list(a = i, b = i^2), mc.cores = 1)
}

