% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mvtest.R
\name{mvtest}
\alias{mvtest}
\title{Permutation test of F == G based on first two moments
To be replaced by Rcpp version}
\usage{
mvtest(x, y, n.reps = 1000, seed = NA, fullOut = FALSE)
}
\arguments{
\item{x}{(n1 x p) data matrix}

\item{y}{(n2 x p) data matrix}

\item{n.reps}{number of replications}
}
\description{
Permutation test of F == G based on first two moments
To be replaced by Rcpp version
}
\examples{
n <- 20
p <- 3
x <- MASS::mvrnorm(n, rep(0, p), diag(rep(1, p)))
y <- MASS::mvrnorm(n, rep(0.1, p), diag(rep(1, p)))
mvtest(x, y, 1e3)
}

