% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mvtest.R
\name{mvtest_stat}
\alias{mvtest_stat}
\title{Permutation test of F == G based on first two moments
To be replaced by Rcpp version}
\usage{
mvtest_stat(x, y)
}
\arguments{
\item{x}{(n1 x p) data matrix}

\item{y}{(n2 x p) data matrix}
}
\description{
Permutation test of F == G based on first two moments
To be replaced by Rcpp version
}

