% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bayes.R
\name{posterior_check}
\alias{posterior_check}
\title{Tests accuracy of posterior distribution}
\usage{
posterior_check(prior_draw, data_draw, posterior_draw, n = 1000, k = 10,
  mc.cores = 1)
}
\arguments{
\item{prior_draw}{function() which produces a parameter vector \code{theta}}

\item{data_draw}{function(theta) which produces data vector \code{x}}

\item{posterior_draw}{function(x) which produces \code{theta_hat}
from the posterior conditional on x and the supplied prior}

\item{n}{Number of (theta, x) samples to draw}

\item{k}{Number of clusters to split the data}

\item{mc.cores}{Number of parallel cores to use}
}
\value{
List with global p-value for equality of posterior
and prior in each of k clusters, as well as generated data and clusters
}
\description{
Given black-box functions \code{prior_draw}, \code{data_draw}, and \code{posterior_draw},
checks that the joint distribution of the prior and data match the joint distribution
of the posterior and data.
Draws a number of (parameter, data, posterior) triplets
and then partitions data into clusters.
Checks that the first two moments of the parameter are consistent with the
first two moments of the posterior within each cluster.
}
\examples{
p <- 3
prior_draw <- function() rnorm(p)
data_draw <- function(theta) theta + rnorm(p)
posterior_draw <- function(x) (x + rnorm(p))/2
res <- posterior_check(prior_draw, data_draw, posterior_draw)
}

